requirements.txt
# =============================================================================
# @file    requirements.txt
# @brief   Python dependencies for CommonPy
# @created 2018-10-23
# @license Please see the file named LICENSE in the project directory
# @website https://github.com/caltechlibrary/commonpy
# =============================================================================

PyYAML               >= 5.3.1
boltons              == 21.0.0
dateparser           == 1.0.0
deprecation          == 2.1.0
h2                   == 3.2.0
httpx                == 0.21.1
humanize             >= 3.9.0
pytest               >= 6.1.2
python_dateutil      == 2.8.2
pywin32              >= 301; sys_platform == 'win32'
sidetrack            >= 2.0.0
tldextract           == 3.0.2
validator-collection == 1.5.0
